[#27] extract function for calendar generation into untis service
